requirements.txt
discord.py
python-dotenv
aiohttp
yt-dlp
PyNaCl
Pillow
davey